// src/app/models/product.ts
export class Product {
    name: string;
    images: string[];
    productCode: string;
    starRating: number;
    price: number;
    quantitity: number;
    condition: string;
    brand: string;
    description: string;
    favourite: boolean;
    feedback: boolean;
    inShoppingCart: boolean;
}